Rename cpg_pipes/pipeline/cli_opts.py
